Created custom exceptions

+++ src/main/java/games/ani/fbsender/service/impl/AnswerServiceImpl.java
@@ -1,5 +1,6 @@
 package games.ani.fbsender.service.impl;
 
+import games.ani.fbsender.exeption.ApiRequestException;
 import games.ani.fbsender.model.Answer;
 import games.ani.fbsender.model.User;
 import games.ani.fbsender.repository.AnswerRepository;
@@ -22,17 +23,25 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public Answer create(Answer answer) {
-        return answerRepository.save(answer);
+        if(answer != null) return answerRepository.save(answer);
+        throw new ApiRequestException("Can`t create empty answer!");
     }
 
     @Override
     public Answer readById(Long id) {
-        return answerRepository.getById(id);
+        return answerRepository.findById(id).orElseThrow(() -> new ApiRequestException(String.format("Can`t find answer with id=%d!",id)));
     }
 
     @Override
     public Answer readByUser(User user) {
-        return answerRepository.getByAuthor(user);
+        if(user != null){
+            Answer answer = answerRepository.getByAuthor(user);
+            if(answer != null){
+                return answerRepository.getByAuthor(user);
+            }
+            throw new ApiRequestException(String.format("Cant find answer with author_id=%d!",user.getId()));
+        }
+        throw new ApiRequestException("User is not exist!");
     }
 
     @Override
@@ -42,7 +51,10 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public void deleteByUser(User user) {
-        answerRepository.deleteAllByAuthor(user);
+        if(user != null){
+            answerRepository.deleteAllByAuthor(user);
+        }
+        throw new ApiRequestException("User is not exist!");
     }
 
     @Override
@@ -52,6 +64,9 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public List<Answer> getAllByUser(User user) {
-        return answerRepository.findAllByAuthor(user);
+        if(user != null){
+            return answerRepository.findAllByAuthor(user);
+        }
+        throw new ApiRequestException("User is not exist!");
     }
 }
